Move utsnameStr to arch dependent file
In some archs, the fields in Utsname are int8 and in others, uint8. See
ztypes_* in https://golang.org/src/syscall/
