+++ elf/kernel_version.go
@@ -25,17 +25,6 @@ import (
 	"syscall"
 )
 
-func utsnameStr(in []int8) string {
-	out := make([]byte, len(in))
-	for i := 0; i < len(in); i++ {
-		if in[i] == 0 {
-			break
-		}
-		out = append(out, byte(in[i]))
-	}
-	return string(out)
-}
-
 var versionRegex = regexp.MustCompile(`^(\d+)\.(\d+).(\d+).*$`)
 
 func kernelVersionFromReleaseString(releaseString string) (int, error) {
